Moved @tryghost/email-events to @tryghost/email-service
fix https://linear.app/ghost/issue/86a4a3e1-a5b4-44f2-a2be-8513d59e0441

- this commit moves the lib code and tests to @tryghost/email-service so we can
  colocate code in one place rather than splitting it out across
  packages, which increases the cognitive load and overhead
