+++ ghost/core/core/server/services/email-suppression-list/MailgunEmailSuppressionList.js
@@ -1,5 +1,5 @@
 const {AbstractEmailSuppressionList, EmailSuppressionData, EmailSuppressedEvent} = require('@tryghost/email-suppression-list');
-const {SpamComplaintEvent, EmailBouncedEvent} = require('@tryghost/email-events');
+const {SpamComplaintEvent, EmailBouncedEvent} = require('@tryghost/email-service');
 const DomainEvents = require('@tryghost/domain-events');
 const logging = require('@tryghost/logging');
 const models = require('../../models');
