+++ ghost/core/test/unit/server/services/members-events/last-seen-at-updater.test.js
@@ -5,7 +5,7 @@ const LastSeenAtUpdater = require('../../../../../core/server/services/members-e
 const DomainEvents = require('@tryghost/domain-events');
 const {MemberPageViewEvent, MemberCommentEvent, MemberSubscribeEvent, MemberLinkClickEvent} = require('@tryghost/member-events');
 const moment = require('moment');
-const {EmailOpenedEvent} = require('@tryghost/email-events');
+const {EmailOpenedEvent} = require('@tryghost/email-service');
 const EventEmitter = require('events');
 const logging = require('@tryghost/logging');
