--- ghost/email-events/index.js
@@ -1,8 +1,0 @@
-module.exports = {
-    SpamComplaintEvent: require('./lib/SpamComplaintEvent'),
-    EmailBouncedEvent: require('./lib/EmailBouncedEvent'),
-    EmailDeliveredEvent: require('./lib/EmailDeliveredEvent'),
-    EmailOpenedEvent: require('./lib/EmailOpenedEvent'),
-    EmailUnsubscribedEvent: require('./lib/EmailUnsubscribedEvent'),
-    EmailTemporaryBouncedEvent: require('./lib/EmailTemporaryBouncedEvent')
-};
