--- ghost/email-events/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@tryghost/email-events",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/email-events",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --check-coverage --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "bson-objectid": "2.0.4",
-    "c8": "8.0.1",
-    "mocha": "10.8.2",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {}
-}
