+++ ghost/email-service/lib/events/EmailDeliveredEvent.js
