+++ ghost/email-service/lib/events/EmailOpenedEvent.js
