+++ ghost/email-service/lib/events/EmailUnsubscribedEvent.js
