+++ ghost/email-service/lib/events/SpamComplaintEvent.js
