+++ ghost/email-service/package.json
@@ -26,7 +26,6 @@
   },
   "dependencies": {
     "@tryghost/color-utils": "0.2.2",
-    "@tryghost/email-events": "0.0.0",
     "@tryghost/errors": "1.3.5",
     "@tryghost/html-to-plaintext": "0.0.0",
     "@tryghost/kg-default-cards": "10.1.1",
