+++ ghost/email-service/test/events/EmailTemporaryBouncedEvent.test.js
@@ -1,6 +1,6 @@
 const assert = require('assert/strict');
 const ObjectID = require('bson-objectid').default;
-const EmailTemporaryBouncedEvent = require('../../lib/EmailTemporaryBouncedEvent');
+const EmailTemporaryBouncedEvent = require('../../lib/events/EmailTemporaryBouncedEvent');
 
 describe('EmailTemporaryBouncedEvent', function () {
     it('exports a static create method to create instances', function () {
