+++ ghost/email-service/test/events/SpamComplaintEvent.test.js
@@ -1,6 +1,6 @@
 const assert = require('assert/strict');
 const ObjectID = require('bson-objectid').default;
-const SpamComplaintEvent = require('../../lib/SpamComplaintEvent');
+const SpamComplaintEvent = require('../../lib/events/SpamComplaintEvent');
 
 describe('SpamComplaintEvent', function () {
     it('exports a static create method to create instances', function () {
@@ -11,5 +11,15 @@ describe('SpamComplaintEvent', function () {
             timestamp: new Date()
         });
         assert(event instanceof SpamComplaintEvent);
+    });
+
+    it('can create an instance without a timestamp', function () {
+        const event = SpamComplaintEvent.create({
+            email: 'test@test.test',
+            memberId: new ObjectID().toHexString(),
+            emailId: new ObjectID().toHexString()
+        });
+        assert(event instanceof SpamComplaintEvent);
+        assert(event.timestamp instanceof Date);
     });
 });
